Allow audience to be passed during create
We may want to pass the `audience` attribute during create instead of setting it via `to` if we're paranoid about the notification being seen during the moment between creation and the subsequent update.

This commit allows this ensuring `private` flag is set when saving such that it agrees with the presence or absence of an audience.